Changes to SPRRZnrc:
Although no change was actually required, a few changes were made to
ensure enough information is provided to user after a parallel run in
the final output file. In many cases, files from individual jobs are
deleted to avoid too many files in one folder. Hence it is desirable
to have calculation information on the final *.egslog and *.egslst
files.

- Add call to ISUMRY before calling egs_combine_runs. Maybe a call
  to routine INPUTS might be more appropriate!

- After combining parallel jobs, jump to new label :TIMING-INFO: from
  where total timing info is generated, followed by the final statistical
  analysis of the combined results.

- Minor change to format of version message to allow for integers of any
  size automatically.

- Removed duplicated verison message from routine INPUTS