Add -u (--urc) option to egs++ apps
For a consistent use of the uniform run control approach across
all EGSnrc apps, introduced the ability to let egs++ apps use a
uniform run control object (URCO) via the command line flag -u
or --urc. This is similar to the use of -s (--simple-run) for a
simple RCO.

If flag -u or --urc passed to the egs++ app on the command line,
a default URCO is created which uses the last job as the one
combining all jobs that have finished and checking for running jobs
for 5 s at 1 s intervals. The user can modify this behavior via the
input file to set other jobs to watch and combine finished jobs, and
the length of the checking interval.

The boolean uniform_run was added to the EGS_Application class for this
purpose.

Differences between -s (simple RCO) and -u (uniform RCO):

- A simple URC is like a URCO that has the last job combine available *egsdat
  files without the ability to wait for a user-defined time interval
  for other jobs to finish
- A uniform RCO allows users to tweak the checking time and define
  which jobs check and combine parallel jobs via input.