Fix bug in JCF object using ncase reduced by njobs
During the implementation of the default URCO request via
the -u switch, a bug was uncovered that cause the JCFO to
run a total number of histories reduced by the number of
parallel jobs. The bug was introduced in the initial
implementation of the URCO where ncase was reduced in
the base class of the RCO's. This works for the simple and
uniform types, but not for the JCFO where the number of histories
are managed by the job control file object (via a lock file).

FIX: Reset ncase to the user-requested value by multiplying it
     by the number of parallel jobs.

Also made sure that by default the last job is a watcher job in the
URCO even if no input block defining an RCO is found in the input
file.