Fix issue: BEAM library fails to rewind lock file.
During the testing phase of the URC approach for Mortran apps,
it was discovered that fort.* files were left behind in the
DOSXYZnrc app folder when using BEAMnrc libraries as a source
of particles (OS: Ubuntu 18.04, compiler: gcc 7.5). Interestingly,
these files are not produced on our IRS systems (CentOS 7, gcc 4.8).
This was regardless of whether a lock-file or an uniform run control
approach was being used.

These fort.* files contained an error message from the BEAM source,
reporting failure to rewind the lock file. The error message was traced
to the egs_pjob_finish routine attemping to rewind and read the lock file.

BUG: The bug is that BEAM sources provide particles to the driving app
     and as such do not create a lock file. But when a parallel run is
     finished, a call from beamlib_finish to beam_finish(ientry) is made
     with ientry = 1. This is the same ientry value used in regular BEAM
     simulations which would indeed create a lock file.

This bug does not have a serious effect on the simulation results as it
just aborts the parallel job when it was already finished with a call to
$egs_fatal which attempts to ouput an error message to the output file,
which by then is closed.

FIX: Call beam_finish from beamlib_finish with a different ientry value
     of 3. This signals that a BEAM source is finished and there is no
     need to call egs_pjo_finish.