Provide information after DOSXYZnrc parallel run
No actual changes were required for the URC implementation.
Added calls to summarizing routines to provide some information
in the combined egslog and egslst files.

In DOSXYZnrc there is only a combine routine implemented for
*.pardose files. As a consequence, the only information available
in the final output files is referent to the total dose in each
requested region of the phantom.

Added calls to routines srcout, and show_transport_parameters
providing more information about the calculation for reference.
This way, if the output files from individual jobs are erased,
for instance for cleanup purposes, this information is still
available.

NOTE: The only information missing in the final output is timing.