Fix pegs4 compilation errors for gcc > 10.0
Since GNU gcc version 10.0 and up, pegs4 compilation fails complaining
about interface mismatch when using dummy arguments for functions. Previous
versions (8.x and 9.x) used to report this as warnings.

Joey Dumont is credited with reporting this error and pointing to a solution based
on INTERFACE blocks. However this wasn't working due to Mortran interpreting :: as
empty labels.

The solution is to use compiler directives %M and %F to switch between Fortran and
Mortran code. Rather than using full INTERFACE blocks, the function type is included
in the external statement for these functions. Subroutines are not affected.

- Add 'clean' target to Makefile

Tested with a few materials that pegs4 data sets created with the modified version were
identical to previous ones.

Maybe it is also needed in the pegsless implementation?